0.0.9.4 jsp pages (login / index/ register) !!! css not working

+++ src/main/java/com/training/app/model/entity/User.java
@@ -211,7 +211,7 @@ public class User {
                 ", lastName='" + lastName + '\'' +
                 ", phoneNumber='" + phoneNumber + '\'' +
                 ", userRole=" + userRole +
-                ", rating=" + rating +
+                ", rating=" + rating.doubleValue() +
                 ", cards=" + cards +
                 ", appointmentList=" + appointmentList +
                 '}';
